requirements.txt
line_bot_sdk>=1.15.0
matplotlib>=3.1.2
Scrapy>=1.7.3
pandas>=0.25.3
pymagnitude>=0.1.120
lxml>=4.4.2
requests>=2.22.0
Twisted>=18.9.0
tensorflow>=1.13.1
yfinance>=0.1.49
tqdm>=4.28.1
jieba>=0.42.1
emoji>=0.5.4
peewee>=3.7.0
Flask>=1.1.1
fnv>=0.2.0
six>=1.14.0
scipy>=1.3.3
gensim>=3.8.1
annoy>=1.16.3
selenium>=3.141.0
numpy>=1.17.4
mpl_finance>=0.10.0
beautifulsoup4>=4.8.2
google_api_python_client>=1.7.11
Pillow>=7.0.0
scikit_learn>=0.22.1
wand
regex